[BE-345] [BE-346] [BE-343] [BE-348] [BE-349]

+++ fooding-core/src/main/java/im/fooding/core/dto/request/menu/MenuCreateRequest.java
@@ -10,7 +10,7 @@ public record MenuCreateRequest(
         Store store,
         MenuCategory category,
         String name,
-        BigDecimal price,
+        int price,
         String description,
         String imageUrl,
         int sortOrder,
